+++ models/genre.js
@@ -1,18 +1,20 @@
 const mongoose = require("mongoose");
-const { Schema } = mongoose;
 
-const GenreSchema = new Schema({
-  name: {
-    type: String,
-    required: true,
-    minlength: 3,
-    maxlength: 100,
-  },
+const Schema = mongoose.Schema;
+
+const genreSchema = new Schema({
+  name: { 
+    type: String, 
+    required: true, 
+    minLength: 3, 
+    maxLength: 100 }
 });
 
-// Virtual for genre's URL
-GenreSchema.virtual("url").get(function () {
-  return `/catalog/genre/${this._id}`;
+// Virtual for Genre's URL
+genreSchema.virtual("url").get(function () {
+  // We don't use an arrow function as we'll need the this object
+  return `/catalog/genre/${this._name}`;
 });
 
-module.exports = mongoose.model("Genre", GenreSchema);
+// Export model
+module.exports = mongoose.model("genre", genreSchema);
